+++ source/extensions/omni.isaac.tests/omni/isaac/tests/tests/test_amr_transporter.py
@@ -19,20 +19,23 @@ import omni.graph.core as og
 import time
 import math
 
-from omni.isaac.core import World
 from omni.isaac.core.utils.nucleus import get_assets_root_path
 from omni.isaac.dynamic_control import _dynamic_control
 from omni.isaac.dynamic_control import utils as dc_utils
-from omni.isaac.core.utils.stage import open_stage_async
 from omni.isaac.core.utils.rotations import quat_to_euler_angles
-from .robot_helpers import init_robot_sim, setup_robot_og
+from omni.isaac.core.utils.extensions import get_extension_path_from_name
+from .robot_helpers import init_robot_sim, setup_robot_og, set_physics_frequency
+from omni.isaac.core import World
+
+from omni.isaac.core.utils.stage import open_stage_async
 from omni.isaac.core.utils.prims import delete_prim
 
 
 # Having a test class dervived from omni.kit.test.AsyncTestCase declared on the root of module will make it auto-discoverable by omni.kit.test
-class TestJetBot(omni.kit.test.AsyncTestCase):
+class TestTransporter(omni.kit.test.AsyncTestCase):
     # Before running each test
     async def setUp(self):
+
         self.dc = _dynamic_control.acquire_dynamic_control_interface()
 
         self._assets_root_path = get_assets_root_path()
@@ -40,21 +43,26 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
             carb.log_error("Could not find Isaac Sim assets folder")
             return
 
-        # add in jetbot (from nucleus)
-        self.usd_path = self._assets_root_path + "/Isaac/Robots/Jetbot/jetbot.usd"
+        self._extension_path = get_extension_path_from_name("omni.isaac.tests")
+
+        ## setup transporter:
+        # open local transporter
+        # (result, error) = await omni.usd.get_context().open_stage_async(self._extension_path + "/data/tests/transporter.usd")
+
+        # add in carter (from nucleus)
+        self.usd_path = self._assets_root_path + "/Isaac/Robots/Transporter/transporter.usd"
         (result, error) = await open_stage_async(self.usd_path)
+
         # Make sure the stage loaded
         self.assertTrue(result)
-
         await omni.kit.app.get_app().next_update_async()
-
         self.my_world = World(stage_units_in_meters=1.0)
         await self.my_world.initialize_simulation_context_async()
 
         # setup omnigraph
         self.graph_path = "/ActionGraph"
         graph, self.odom_node = setup_robot_og(
-            self.graph_path, "left_wheel_joint", "right_wheel_joint", "/jetbot", 0.0335, 0.118
+            self.graph_path, "left_wheel_joint", "right_wheel_joint", "/Transporter", 0.08, 0.58
         )
 
         pass
@@ -67,10 +75,8 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         # In some cases the test will end before the asset is loaded, in this case wait for assets to load
         while omni.usd.get_context().get_stage_loading_status()[2] > 0:
             await omni.kit.app.get_app().next_update_async()
-
         pass
 
-    # Actual test, notice it is "async" function, so "await" can be used if needed
     async def test_loading(self):
 
         delete_prim("/ActionGraph")
@@ -79,7 +85,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
 
         # get the dofbot
-        self.ar = self.dc.get_articulation("/jetbot")
+        self.ar = self.dc.get_articulation("/Transporter")
         self.chassis = self.dc.get_articulation_root_body(self.ar)
         self.starting_pos = np.array(self.dc.get_rigid_body_pose(self.chassis).p)
 
@@ -110,7 +116,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/jetbot")
+        await init_robot_sim(self.dc, "/Transporter")
 
         for x in range(1, 5):
             forward_velocity = x * 0.15
@@ -141,7 +147,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/jetbot")
+        await init_robot_sim(self.dc, "/Transporter")
         for x in range(1, 5):
             self.my_world.play()
             await omni.kit.app.get_app().next_update_async()
@@ -157,7 +163,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
                 await omni.kit.app.get_app().next_update_async()
             og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(0.0)
             og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:angularVelocity").set(0.0)
-            for j in range(10):
+            for j in range(30):
                 await omni.kit.app.get_app().next_update_async()
             self.assertAlmostEqual(og.DataView.get(odom_velocity)[0], 0.0, delta=5e-1)
             self.assertAlmostEqual(og.DataView.get(odom_ang_vel)[2], 0.0, delta=5e-1)
@@ -168,27 +174,25 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
 
     async def test_spin(self):
         odom_ang_vel = og.Controller.attribute("outputs:angularVelocity", self.odom_node)
+        # Start Simulation and wait
+        self.my_world.play()
+        await omni.kit.app.get_app().next_update_async()
+        await init_robot_sim(self.dc, "/Transporter")
 
-        for x in range(1, 6):
-            # Start Simulation and wait
-            self.my_world.play()
-            await omni.kit.app.get_app().next_update_async()
-
-            await init_robot_sim(self.dc, "/jetbot")
-
+        for x in range(1, 4):
             angular_velocity = 0.6 * x
             og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:angularVelocity").set(
                 angular_velocity
             )
 
             # wait until const velocity reached
-            for i in range(30):
+            for i in range(60):
                 await omni.kit.app.get_app().next_update_async()
 
             curr_ang_vel = float(og.DataView.get(odom_ang_vel)[2])
             self.assertAlmostEqual(curr_ang_vel, angular_velocity, delta=5e-2)
 
-        self.my_world.stop()
+        # self.my_world.stop()
 
         pass
 
@@ -202,7 +206,7 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         self.my_world.play()
         await omni.kit.app.get_app().next_update_async()
 
-        await init_robot_sim(self.dc, "/jetbot")
+        await init_robot_sim(self.dc, "/Transporter")
         forward_velocity = -0.1
         angular_velocity = -0.5
         og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(forward_velocity)
@@ -219,69 +223,3 @@ class TestJetBot(omni.kit.test.AsyncTestCase):
         await omni.kit.app.get_app().next_update_async()
 
         pass
-
-    # corner case: quick accel from zero
-    # async def test_jetbot_accel_drop_reset(self):
-    #     odom_velocity = og.Controller.attribute("outputs:linearVelocity", self.odom_node)
-    #     odom_ang_vel = og.Controller.attribute("outputs:angularVelocity", self.odom_node)
-
-    #     # Start Simulation and wait
-    #     self.my_world.play()
-    #     await omni.kit.app.get_app().next_update_async()
-
-    #     await init_robot_sim(self.dc, "/jetbot")
-    #     l_wheel = self.dc.get_rigid_body("/jetbot/left_wheel")
-
-    #     forward_velocity = 0.2
-    #     og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(forward_velocity)
-
-    #     # wait until const velocity reached
-    #     for i in range(50):
-    #         await omni.kit.app.get_app().next_update_async()
-
-    #     curr_t = 0
-    #     for i in range(1600):
-    #         if i - curr_t >= 200:
-    #             self.my_world.stop()
-    #             og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(0)
-    #             await omni.kit.app.get_app().next_update_async()
-    #             curr_t = i
-    #             self.my_world.play()
-    #             await omni.kit.app.get_app().next_update_async()
-
-    #             await init_robot_sim(self.dc, "/jetbot")
-    #             # l_wheel = self.dc.get_rigid_body("/jetbot/left_wheel")
-
-    #             # wait until const velocity reached
-    #             for j in range(100):
-    #                 await omni.kit.app.get_app().next_update_async()
-
-    #             forward_velocity += 0.2
-    #             og.Controller.attribute(self.graph_path + "/DifferentialController.inputs:linearVelocity").set(
-    #                 forward_velocity
-    #             )
-
-    #             print("forward velocity: " + str(forward_velocity))
-
-    #             # wait until const velocity reached
-    #             for j in range(100):
-    #                 await omni.kit.app.get_app().next_update_async()
-
-    #         if og.DataView.get(odom_ang_vel)[2] > 0.8:
-    #             print("spinning out of control!")
-    #             print("linear velocity: " + str(forward_velocity))
-    #             self.my_world.stop()
-
-    #         else:
-    #             curr_vel = float(og.DataView.get(odom_velocity)[0])
-    #             self.assertAlmostEqual(curr_vel, forward_velocity, delta=5e-1)
-    #             self.assertAlmostEqual(
-    #                 curr_vel, (self.dc.get_rigid_body_angular_velocity(l_wheel)[1]) * 0.0325, delta=0.25
-    #             )
-    #             print(self.dc.get_rigid_body_angular_velocity(l_wheel)[1] * 0.0325)
-    #             print("correct forward velocity: " + str(forward_velocity))
-    #         await omni.kit.app.get_app().next_update_async()
-
-    #     self.my_world.stop()
-
-    #     pass